{{cookiecutter.repo_name}}/{{cookiecutter.project_name}}/api/tests.py
from rest_framework import status
from rest_framework.test import APITestCase


class SampleTestCase(APITestCase):
    def test_url_root(self):
        url = '/api/'
        response = self.client.get(url)
        self.assertTrue(status.is_success(response.status_code))
